% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqi_cities.R, R/aqi_cities_imp.R
\docType{data}
\name{aqi_cities}
\alias{aqi_cities}
\title{Air Quality Index per City Example and Test Dataset}
\format{
An object of class \code{"data.frame"}.

An object of class \code{"data.frame"}.
}
\usage{
data(aqi_cities)

data(aqi_cities_imp)
}
\description{
Air Quality Index per City Example and Test Dataset

Air Quality Index per City Example and Test Dataset - Imputed with Mean Values for Testing
}
\keyword{datasets}
