% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aqi_cities_t.R
\docType{data}
\name{aqi_cities_t}
\alias{aqi_cities_t}
\title{Air Quality Index per City Example and Test Dataset, Transposed}
\format{
An object of class \code{"data.frame"}.
}
\usage{
data(aqi_cities_t)
}
\description{
Air Quality Index per City Example and Test Dataset, Transposed
}
\keyword{datasets}
