% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareSummary.R
\name{compareSummary}
\alias{compareSummary}
\title{Compare Summaries of Original vs. Imputed Data}
\usage{
compareSummary(original, imputed, by)
}
\arguments{
\item{original}{The original, unimputed data object.}

\item{imputed}{The data object with missing values imputed.}
}
\value{
Returns a data frame containing the mean, median, and standard
deviation of both data objects.
}
\description{
Compare Summaries of Original vs. Imputed Data
}
\examples{
compareSummary(data, data_imputed)
}
