% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputationHotDeck.R
\name{imputationHotDeck}
\alias{imputationHotDeck}
\title{Hot Deck Imputation}
\usage{
imputationHotDeck(
  x,
  by.row = FALSE,
  row.range = 1:nrow(x),
  col.range = 1:ncol(x)
)
}
\arguments{
\item{x}{A matrix-like R object with two dimensions to perform imputation on.}

\item{by.row}{A logical value indicating whether to calculate the mean by the
values within an element's row or not. Default = \code{TRUE}}

\item{row.range}{A numeric vector containing the row numbers to be included
in the hot deck imputation.}

\item{col.range}{A numeric vector containing the column numbers to be
included in the hot deck imputation.}
}
\value{
Returns the passed data with missing values imputed with a random
value from that element's column or row.
}
\description{
Replace missing values with a randomly selected, non-NA value from that
element's column or row.
}
\examples{
# Default by column
imputed_mean <- imputationMean(data)

# By row, excluding the first column.
imputed_mean_byrow <- imputationMean(data_byrow,
col.range = 2:ncol(data_byrow))
}
